rename source directory
