+++ pytest_gpiozero/devices.py
