+++ pytest_gpiozero/fonts/7seg.txt
