+++ pytest_gpiozero/input_devices.py
