+++ pytest_gpiozero/internal_devices.py
