+++ pytest_gpiozero/mixins.py
