+++ pytest_gpiozero/output_devices.py
