+++ pytest_gpiozero/pins/native.py
