+++ pytest_gpiozero/pins/spi.py
