+++ pytest_gpiozero/spi_devices.py
