+++ pytest_gpiozero/tools.py
