+++ setup.cfg
@@ -48,7 +48,7 @@ test =
 testpaths = tests
 
 [coverage:run]
-source = gpiozero
+source = pytest_gpiozero
 branch = true
 
 [coverage:report]
